+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+2016/07/13: v0.0.3
+==================
+
+* RDS type system with BLAS compatible complex type and Cast trait.
+* Several array manipulation functions: insert, extract, remove, split.
+* EDF format reader.
+
 2016/06/20: v0.0.2
-=================
+==================
 
 * NDArray indexing, slicing, transpose.
 * Numpy array and CSV file format support.